+++ src/s17/taskmanager/subscribers.py
@@ -5,7 +5,7 @@ from five import grok
 from Products.CMFCore.interfaces import IActionSucceededEvent
 from s17.taskmanager.adapters import IResponse
 from s17.taskmanager.content import ITask
-from s17.taskmanager.content import ITaskFolder
+from s17.taskmanager.content import ITaskPanel
 from zope.lifecycleevent.interfaces import IObjectAddedEvent
 from zope.lifecycleevent.interfaces import IObjectModifiedEvent
 
@@ -70,7 +70,7 @@ def set_task_end_date(task, event):
         task.end_date = None
 
 
-@grok.subscribe(ITaskFolder, IObjectAddedEvent)
+@grok.subscribe(ITaskPanel, IObjectAddedEvent)
 def set_add_taskfolder_local_role(taskfolder, event):
     ''' This subscriber will create a local role enable responsible to
         add tasks
@@ -82,7 +82,7 @@ def set_add_taskfolder_local_role(taskfolder, event):
         taskfolder.manage_setLocalRoles(taskfolder.responsible, ['Manager'],)
 
 
-@grok.subscribe(ITaskFolder, IObjectModifiedEvent)
+@grok.subscribe(ITaskPanel, IObjectModifiedEvent)
 def set_edit_taskfolder_local_role(taskfolder, event):
     ''' This subscriber will create a local role enable responsible to
         add tasks