+++ src/s17/taskmanager/tests/test_taskpanel.py
@@ -5,7 +5,7 @@ from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.uuid.interfaces import IAttributeUUID
-from s17.taskmanager.content import ITaskFolder
+from s17.taskmanager.content import ITaskPanel
 from s17.taskmanager.testing import INTEGRATION_TESTING
 from zope.component import createObject
 from zope.component import queryUtility
@@ -30,7 +30,7 @@ class IntegrationTest(unittest.TestCase):
         self.obj = self.folder['obj']
 
     def test_adding(self):
-        self.assertTrue(ITaskFolder.providedBy(self.obj))
+        self.assertTrue(ITaskPanel.providedBy(self.obj))
 
     def test_fti(self):
         fti = queryUtility(IDexterityFTI, name=ctype)
@@ -39,13 +39,13 @@ class IntegrationTest(unittest.TestCase):
     def test_schema(self):
         fti = queryUtility(IDexterityFTI, name=ctype)
         schema = fti.lookupSchema()
-        self.assertEqual(ITaskFolder, schema)
+        self.assertEqual(ITaskPanel, schema)
 
     def test_factory(self):
         fti = queryUtility(IDexterityFTI, name=ctype)
         factory = fti.factory
         new_object = createObject(factory)
-        self.assertTrue(ITaskFolder.providedBy(new_object))
+        self.assertTrue(ITaskPanel.providedBy(new_object))
 
     def test_is_referenceable(self):
         self.assertTrue(IReferenceable.providedBy(self.obj))