+++ src/s17/taskmanager/upgrades/to1000.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-def fromZero(context):
-    """ Upgrade from Zero to version 1000
-    """